meld-captioning-pipeline/requirements.txt
transformers
whisper
accelerate
timm
torchvision
torchaudio
torch
pandas
tqdm
opencv-python
# cv2, if with conda, run "conda install -c conda-forge opencv"
soundfile
scipy
pillow
